Round BJ payouts up to nearest integer; TotalWon and BankNet are now long
Math.Ceiling replaces Math.Round(..., 2) so odd bets (e.g. 1 gil at 3:2)
yield whole-number payouts. PlayerStat.TotalWon and RoundHistoryEntry.BankNet
changed from decimal to long to enforce integer-only gil amounts.

+++ TwentyOne.Game/GameTypes.cs
@@ -9,7 +9,7 @@ public class RoundHistoryEntry
     public int       RoundNumber { get; set; }
     public GameState Snapshot    { get; set; } = new();
     // Bank net from the venue's perspective: positive = bank gained, negative = bank paid out.
-    public decimal   BankNet     { get; set; }
+    public long      BankNet     { get; set; }
     // Player bank balances after payout (key = PlayerStatKey). Used by history window tooltips.
     public Dictionary<string, long> PlayerBanks { get; set; } = [];
 }
